requirements.txt
django==4.2.7
pillow==10.1.0
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
face-recognition==1.3.0
numpy==1.26.2
dlib==19.24.2
django-extensions==3.2.3
werkzeug==2.3.7
pyOpenSSL==23.2.0
python-magic==0.4.27
django-crispy-forms==2.1
crispy-tailwind==0.5.0
python-docx==0.8.11